# Improve `diag_At_A` and `diag_At_B` for vectors