```julia
julia> using BenchmarkTools

julia> x = rand(5000); y = rand(5000);

# before
julia> @btime vec(sum(abs2.(x); dims=1));
  2.800 μs (6 allocations: 39.23 KiB)

julia> @btime vec(sum(x .* y; dims=1));
  2.391 μs (5 allocations: 39.23 KiB)

# after
julia> @btime [x'x];
  515.724 ns (3 allocations: 96 bytes)

julia> @btime [x'y];
  813.826 ns (3 allocations: 96 bytes)
```