@sharanry 2022-03-21
@devmotion Is this good to merge?
I have :
- Bumped the patch version
- Added allocation tests
- Updated the initial post